+++ forui/lib/src/widgets/select/basic_select.dart
@@ -42,6 +42,7 @@ class _BasicSelect<T> extends FSelect<T> {
     super.contentScrollController,
     super.contentScrollHandles,
     super.contentPhysics,
+    super.divider,
     super.initialValue,
     super.key,
   }) : super._();
@@ -64,6 +65,7 @@ class _BasicSelectState<T> extends _State<_BasicSelect<T>, T> {
       enabled: widget.enabled,
       scrollHandles: widget.contentScrollHandles,
       physics: widget.contentPhysics,
+      divider: widget.divider,
       children: widget.children,
     );
   }